Removed the need for two separate subprojects
- integrated the ZooKeeper Schema Exposer into the main project;
- upgraded most dependencies & sbt plugins to newer versions;
