+++ .scalafmt.conf
@@ -1,0 +1,1 @@
+version = 2.6.4
