--- kafkatime/src/test/scala/org/codefeedr/kafkatime/transforms/SchemaConverterTest.scala
@@ -1,84 +1,0 @@
-package org.codefeedr.kafkatime.transforms
-
-import java.nio.ByteBuffer
-
-import com.sksamuel.avro4s._
-import org.apache.avro.Schema
-import org.apache.avro.io.FastReaderBuilder.RecordReader.Stage
-import org.codefeedr.kafkatime.transforms.SchemaConverter.getNestedSchema
-import org.scalatest.funsuite.AnyFunSuite
-import org.scalatest.prop.{TableDrivenPropertyChecks, TableFor2}
-
-import scala.language.higherKinds
-
-class SchemaConverterTest extends AnyFunSuite with TableDrivenPropertyChecks {
-
-  case class Nested(someValue: Int)
-
-  case class someStage(someStage: Stage)
-
-  case class AllSupportedTypes(someString: String, someFloat: Float, someDouble: Double, someInt: Int,
-                               someBoolean: Boolean, someLong: Long, someOptional: Option[String], someByte: ByteBuffer,
-                               someMap: Map[String, Int], someArray: Array[Int], someList: List[Long], someNested: Nested,
-                               someNestedList: List[Nested], someNull: Null)
-
-  def getSchema: Schema = {
-    implicit object ItemToSchema extends SchemaFor[Null] {
-      override def schema(fieldMapper: FieldMapper): Schema = {
-        Schema.create(Schema.Type.NULL)
-      }
-    }
-    implicit object NullEncoder extends Encoder[Null] {
-      override def encode(t: Null,
-                          schema: Schema,
-                          fieldMapper: FieldMapper): AnyRef = {
-        t
-      }
-    }
-    implicit object DateDecoder extends Decoder[Null] {
-      override def decode(value: Any,
-                          schema: Schema,
-                          fieldMapper: FieldMapper): Null = {
-        null
-      }
-    }
-    AvroSchema[AllSupportedTypes]
-  }
-  val schema: Schema = getSchema
-
-  val testData: TableFor2[String, String] =
-    Table(
-      ("expectedType", "FieldName"),
-      ("STRING", "someString"),
-      ("FLOAT", "someFloat"),
-      ("DOUBLE", "someDouble"),
-      ("INTEGER", "someInt"),
-      ("BOOLEAN", "someBoolean"),
-      ("BIGINT", "someLong"),
-      ("STRING", "someOptional"),
-      ("BYTES", "someByte"),
-      ("MAP<STRING, INTEGER>", "someMap"),
-      ("ARRAY<INTEGER>", "someArray"),
-      ("ARRAY<BIGINT>", "someList"),
-      ("ROW<someValue INTEGER>", "someNested"),
-      ("ARRAY<ROW<someValue INTEGER>>", "someNestedList"),
-      ("NULL", "someNull")
-    )
-
-  /**
-   * Parameterized good weather tests for all supported types.
-   */
-  forAll(testData) { (t: String, name: String) =>
-    assertResult((name, t)) {
-      val res = getNestedSchema(name, schema.getField(name).schema())
-      (res._1, res._2.toString)
-    }
-  }
-  /**
-   * Unsupported type.
-   */
-  assertThrows[RuntimeException] {
-    getNestedSchema("doesnt matter", AvroSchema[someStage])
-  }
-
-}
